+++ primefaces/src/main/java/org/primefaces/component/tree/TreeComponentHandler.java
@@ -23,6 +23,7 @@
  */
 package org.primefaces.component.tree;
 
+import java.util.Locale;
 import javax.faces.view.facelets.ComponentConfig;
 import javax.faces.view.facelets.ComponentHandler;
 import javax.faces.view.facelets.MetaRule;
@@ -35,6 +36,9 @@ public class TreeComponentHandler extends ComponentHandler {
     private static final MetaRule DROP_LISTENER
             = new MethodRule("onDrop", Boolean.class, new Class[]{TreeDragDropInfo.class});
 
+    private static final MetaRule FILTER_FUNCTION
+            = new MethodRule("filterFunction", Boolean.class, new Class[]{Object.class, Object.class, Locale.class});
+
     public TreeComponentHandler(ComponentConfig config) {
         super(config);
     }
@@ -45,6 +49,7 @@ public class TreeComponentHandler extends ComponentHandler {
         MetaRuleset metaRuleset = super.createMetaRuleset(type);
 
         metaRuleset.addRule(DROP_LISTENER);
+        metaRuleset.addRule(FILTER_FUNCTION);
 
         return metaRuleset;
     }